Moving showRouteFromCurrentPosition to showRoute.js
