+++ app/javascript/packs/show_route_from_current_position.js
@@ -1,3 +1,3 @@
-import { showRouteFromCurrentPosition } from '../components/showRouteFromCurrentPosition';
+import { showRouteFromCurrentPosition } from '../components/showRoute';
 
 showRouteFromCurrentPosition();
